// examples/cross_canister_calls/test/dfx_generated/canister1/canister1.did.d.ts
import type { Principal } from '@dfinity/principal';
import type { ActorMethod } from '@dfinity/agent';

export interface Account { 'id' : string, 'balance' : bigint }
export interface AccountArgs { 'id' : string }
export type AccountResult = { 'ok' : [] | [Account] } |
  { 'err' : string };
export type AccountsResult = { 'ok' : Array<Account> } |
  { 'err' : string };
export type BalanceResult = { 'ok' : bigint } |
  { 'err' : string };
export type NotifyResult = { 'ok' : null } |
  { 'err' : string };
export type TrapResult = { 'ok' : string } |
  { 'err' : string };
export interface _SERVICE {
  'account' : ActorMethod<[AccountArgs], AccountResult>,
  'accounts' : ActorMethod<[], AccountsResult>,
  'balance' : ActorMethod<[string], BalanceResult>,
  'send_notification' : ActorMethod<[], NotifyResult>,
  'transfer' : ActorMethod<[string, string, bigint], BalanceResult>,
  'trap' : ActorMethod<[], TrapResult>,
}